requirements.txt
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.20.0
